+++ Server/AntiCheatServer.csproj
@@ -55,7 +55,11 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Patch.cs" />
+    <Compile Include="Framework\Patch.cs" />
+    <Compile Include="Framework\PatchDescriptor.cs" />
+    <Compile Include="Framework\Patches\Multiplayer_ParseServerToClientsMessage_Patcher.cs" />
+    <Compile Include="Framework\Patches\Multiplayer_ProcessIncomingMessage_Patcher.cs" />
+    <Compile Include="Framework\Patches\Server_SendMessage_Patcher.cs" />
     <Compile Include="ModEntry.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
